[FIX] l10n_es_reports: fix data upgrade
Fine-tuning of commit 83bd5aa347b.
While it was necessary to fix the report, it was incomplete as it leaved
an inconsistency between the record name which was inverted with the function.
Furthermore it broke the module upgrade since the inversion of the codes
made it so that data upgrade would violate the unicity constraint on codes.

We add a technical device to fix the module upgrade.
Since it is made necessary only by the sorry situation introduced by the
fixed oversight, it could be removed in master.

opw 1954820