[ADD] l10n_mx_edi_reports_closing: Option to get closing entry was added
We at Mexico need to report the closing entry (A.K.A Month 13th move) in
the trial balance.

https://www.sat.gob.mx/articulo/97859/regla-2.8.1.7

This **especial** report must consider all the accounts affected in the
fiscal year, and only the balance accounts in the closing entry must
have moves.

We solved it by adding to the account move a boolean field to indicate
which is the closing entry for the year. We create a new report : Trail
Balance Month 13 (inherited from the "normal" trial balance).

We did it in. a new module due to it is necessary a new field to
indicate which are the closing entry.

closes odoo/enterprise#4337

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>


Co-authored-by: Luis Torres <luis_t@vauxoo.com>