[FIX] account_reports: Date format error with Next Reminder Date
Steps to reproduce the bug
- Set the lang of the user to en_US
- Go to the followup report
- Change the date format of the user's lang to %d/%m/%Y
- Choose a partner P with a need of action with Next reminder date dd > 12
- Click on "Print letter" and afterwards click on "Done"

Bug:
A traceback was raised because the format of the date was expected to be in UTC instead
of checking the format date set on the user.

So if the date was 22/05/2019 with isUTC set to true, it considered 22 as the month instead
of the day.

PS: Now the datePattern is checked in function parseDate defined in field_utils.js

opw:1999088

closes odoo/enterprise#4396

Signed-off-by: Simon Goffin (sig) <sig@openerp.com>