[FIX] delivery_bpost: correct shiping value
The shipping value was bad calculated in bpost, because it must be sent
in 'centime' and we take the price in full unit which leads to a
shipping value divided by 100.

More over, some kind of shipping method shouldn't have commercial
values, 'DOCUMENTS', 'GIFT' and 'SAMPLE' so we put the lower value
possible which is 100 centime (1€).

We are also computing the value on stock move lines instead of the total
of sale order, because services, delivery_fees, etc should not be
included in shipping price. More over we should be allowed to get a
value wihtout sale order.

closes odoo/enterprise#4397

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>