[FIX] iot: stopPolling not aborting requests
When adding a listener to the longpolling, we call `stopPolling` to
abort all requests currently working. Then, we start a new request with
`startPolling`.

`stopPolling` was not working as expected and previous requests were
not killed, resulting in a huge number of useless requests still
running.

The problem was that, when we got an answer for the current request, we
first sent the new request before removing the reference to the old
one. Instead of setting the reference to the old request to `false`, we
overrode the reference to the new one, that was still running. The
stopPolling method was then not able to kill the request since it
didn't have its reference.

closes odoo/enterprise#4415

Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>