[FIX] account_invoice_extract: fix rounding issues on taxes
Sometimes, the rounding method used for taxes isn't the same in Odoo than
what is used inside the invoice parsed by the OCR (one is rounding per
invoice line, the other one rounds on the sum).
This commit introduces an adjustment to the taxes lines so that they match
every time.