[FIX] stock_barcode: lot/sn modal on product scan
Since 12.0's odoo/odoo@76662a0a6 stock.picking move_line_ids field is
replaced by move_line_ids_without_package in the form view (which is a
view over move_line_ids field).

The widget picking_barcode_handler show eg. "Lot/Serial Number Details"
modal when barcode of a product in the picking is scanned. But since it
was not adapted to the field change, it did not work since it used the
"move_line_ids" field not in the view.

opw-1997625
closes #4422

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>