[FIX] mrp_workorder: cancel MTO
- Install MRP II and Purchase
- In Manufacturing, activate Multi-Step Routes
- In a warehouse (YourCompany), change Incoming Shipments to
"Receive [...] (2 steps)"
- Create product A
- Create product B, with a vendor, Make To Order, Buy, Tracking By Lots
- Create a Bill of Material for A, with component B
Choose a routing, and component B is consumed in operation
- Create a Manufacturing Order with BOM for A, plan
- It creates a picking, cancel it
- In the MO, click on work order and process.

A traceback is raised on `moves[0]`.

In an ideal case, the corresponding stock move should not be canceled,
but it needs to be handled in master.

We make the code more robust and raise a proper error message with a
workaround.

opw-1986723

closes odoo/enterprise#4424

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>