[FIX] account_budget: fix read_group override for pivot view
The pivot view displays the theoritical_amount (sic), practical amount,
and percentage fields.
Since these fields are computed, the read_group is overriden to add the fields
to the result.
However the 'total' line is wrong on these fields:
it always gives the sum of all budgets lines, regardless of filters
(unlike the practical_amount which takes the filters into account).

The cause is that __domain is not set when there is no groupby
(as in the case of the total line), and so the domain was ignored.

opw 1986655