[FIX] stock_barcode: barcode add notracking product
If a product without tracking is added the first time on a normal
stock.picking form view, there is an added move_line_ids, but since the
field in the view is now move_line_ids_without_package (since commit
odoo/odoo@76662a0a) the line is not added.

With this changeset, the onchange is coherant with the current
expectations, once saved both move_line_ids and
move_line_ids_without_package refer to the same data (with
move_line_ids_without_package that has an additional filtering domain)
so the result should be the same.

opw-1997625
closes #4455

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>