[FIX] stock_barcode: form view and barcode_scanned events
It is possible to open a form view inside the stock barcode client action.
Once this is opened, one would think scanning barcode would be ignored.
This was not the case.

Exemple to reproduce the issue:
* add a new line (it opens the form view of a sml)
* lose the focus
* scan source1 product1 dest1
* discard the form view
a move line of product1 from source1 to dest1 was added silently

We now correctly stop listening to barcodes while in the client action.

task-1974634

closes odoo/enterprise#4502

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>