[FIX] l10n_mx_edi: Allow remove invoice PDF document after validation
Right now if you try to remove the invoice PDF document generated after
validating the invoice an error message will pop up:

```
You cannot delete a set of documents which has a legal information and
it's declared to the SAT, please try to cancel the document linked to
the record: 52 in the model: account.invoice and named:
INV-INV20190003-MX-Invoice-3-3.xml with the UUID:
EDF87330-1E3C-4A2A-AD23-2483E9D7C8B5 first and then resign it if
necessary.
```

This should be only true form XML invoice documents which are the real
legal document, given the fact that you can generate as many PDFs as you
want.

Steps to reproduce:

- Create and Invoice and Validate it
- Print a PDF
- Try to remove the PDF
- The error message previously commented is displayed

Closes #4477

closes odoo/enterprise#4479

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>