[FIX] web_studio: display reports not bound to actions
In studio > report > report tab.
untick the "Add in the print menu" checkbox

Return on the report kanban listing

Before this commit, the report on which we deactivated the action was not present
This was because the domain which searches for reports did not take this
case into account

After this commit, the modified report shows up on the listing.

OPW 2005875

closes odoo/enterprise#4493

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>