[FIX] account_report: cash basis method lack account moves in report
Before this commit, when 'Cash Basis Method' was selected in an account
report, only the moves with journal of kind 'cash' or 'bank' was used in
the report. This is an issue because the report lacks of different
account moves as for example the ones of customer invoices.

Now, all the moves are correctly included.

Technical Note:
---------------

We remove the domain from the _query_get to avoid testing the
user_type_id by the mean of the where clause in the query:
"account_move_line"."account_id" in (%s,%s,%s). The user_type_id it's
already taken into account in a different way, in :
https://github.com/odoo/enterprise/blob/a3d04f2bd13387760f0996066a3a895e38ce0787/account_reports/models/account_financial_report.py#L732

opw-2002310

closes odoo/enterprise#4497

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>