[FIX] account_reports: prevent long line to overflow
Create a payment with a loooooooooooooooooooooong memo and a big amount
(6874118.11$).
Create a bank statement with a loooooooooooooooooooooong name.

Before this commit, the HTML reports displayed a crop of the long line
but still with a horizontal scrollbar.

After this commit, we augmented the force crop of the lines.

Fine-tunning : 66cd71dc804973f43c454fbf359f713921ab703c and be0bedc958fb68a8b686a6929184bbbe3072b82b

closes odoo/enterprise#4506

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>