[FIX] l10n_mx_edi, l10n_mx_edi_payment: avoid missing the partner bank account value in the payment XML generated when changed from bank statement
When a bank statement is generated and the field l10n_mx_edi_partner_bank_id
are set from here, the XML generated for the payment does not consider such
value, because the process_reconciliation method is executed and the payment
XML is generated before the write is done in the payment of such field.

Right now the payment have to be cancel and confirmed again so the XML
is re-generated and take the changes made in the corresponding fields.

Without this change:
https://drive.google.com/file/d/12MY6qQbbbAbFE3Y8vBHYEnC4o8Is2bjp/view

With this change:

The payment does not have to be cancel and re-confirmed.
(Video: https://drive.google.com/file/d/1wUTD5iiuXcrpO3rfycHQKkAjB_HlmUE7/view)

closes odoo/enterprise#4420

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>