[FIX] account_batch_payment, account_sepa_direct_debit: fix fields attrs definitions
Bad definition of fields. From the ORM documentation:
states – a dictionary mapping state values to lists of UI attribute-value pairs
The /web/export/get_fields route makes the hypothesis that this convention
is respected, otherwise it might crash.

This makes it impossible to use import-compatible export on the model.

opw 2003110

closes odoo/enterprise#4510

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>