[FIX] sale_ebay: replace deprecated api.one from stock methods by api.multi
Changing the api.multi of a method into api.one implies a change of behaviour.
There is absolutely no reason for such a thing to happen; the override of stock
methods in this module exist only to add a synchronisation, so it should not
touch the api. This change persisted because it was simply never updated.

opw 2004727

closes odoo/enterprise#4515

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>