[FIX] web_studio: correctly apply store attribute on new related field
Before this commit, when adding a new related field from the studio, it was
always stored by default regardless of the original field.

This commit fixes the issue by adding a proper 'store' attribute while adding
a related field. The attribute will be the same as the field which is selected
on the dialog. And thus it will make the behavior consistent / predictable.

task-1985710

Backport of 6ded01ee6df8ca12ed848501cd0871b477675763

closes odoo/enterprise#5220

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>