[FIX] helpdesk: fix mail templates language
Because of the whitespace, the lang would be e.g. " fr_BE", which is not a valid
language in Odoo; as a result the template would always be rendered in English.

In this commit we learn that:
- copy-paste is evil
- significant whitespace is treacherous
There is a moral to the story as bugs are eventually killed (and go to hell).

opw 2008474

closes odoo/enterprise#4568

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>