[FIX] delivery_bpost: display correct partner name
When a partner which is a shipping address is selected on a picking that
must be sent through bpost integration, you get an error from bpost,
because the name of the partner is empty.

So instead of using the name, we'll use the display_name that will show
the company name followed by 'Shipping address'.

closes odoo/enterprise#4577

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>