[FIX] stock_barcode: scan SN
- Create a product A tracked by SN, set a barcode
- Create a picking with A
- Go to Barcode > Operations and open the picking
- Scan the barcode => the "Wizard to scan SN/LN for specific product"
opens
- Set a quantity done
- Validate Lot

The quantity done is not saved on the picking.

This is because the ORM is not yet ready for a related field without any
relation, as defined on the field `product_barcode`. At creation of the
transient model, `lot_name` is filled in, but `product_barcode` is
empty. Because of the `related` attribute the value of `lot_name` is
set to `False`, then the lot is skipped at validation due to this
condition being falsy:

https://github.com/odoo/enterprise/blob/9b9dc9e1123d4e8db3dd1ea6b047667d7b44dbe3/stock_barcode/wizard/stock_barcode_lot.py#L73

Using a computed field should be the way to go.

opw-2006392

closes odoo/enterprise#4585

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>