[FIX] web_studio: Field Column table w/o t-foreach
In the report editor, when a "Field Column" widget is added in a table
without t-foreach only the title is set. This may be unexpected in the
case of table just displaying one row of data and not looped over.

With this changeset, if there is no t-foreach arround target elements,
we use another behavior of adding the field value inside line that are
not inside a t-foreach.

Without change, the added test fails with:

- "Should select the column (1)" with the table cell that is empty
instead of containing value (`<span t-field="o.child.name"/>`)

opw-2003119
closes #4609

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>