[FIX] web_dashboard: pie chart structure not modified by favorites
A pie chart widget aims to instantiate a graph view in mode
'pie', with a measure and a single groupby specified via some
attributes. It turns out that if a pie chart is instantiated
in a context in which the keys 'graph_mode', 'graph_measure',
and 'graph_groupbys' are present, that keys prevail over the
pie chart specification. This can produce a line chart for
instance. This commit brings a test covering the fix in
community https://github.com/odoo/odoo/pull/34234/commits/171f736c39e19bef7bb004f299a4b41e1587c11f
(historically, the pie chart widgets are tested in dashboard
views).

closes odoo/enterprise#4615

Signed-off-by: VincentSchippefilt <VincentSchippefilt@users.noreply.github.com>