[FIX] account_ponto: fix error when manually refreshing too fast
Ponto use a route in order to synchronize the latest account balance and transactions. The route is the same except for the parameters.
Previously when clicking manual refresh we were calling the route one after another in order to get the new balance and the latest transactions. However doing so returns a HTTP 403 error, and more generally, calling the same synchronization route too fast cause the same error.
Therefore we removed the synchronization of the account balance (as the balance of the account is hidden in a technical menu) and we only refresh the latest transactions. We also add a check in order to do that synchronization only once every 5 minutes.

This commit breaks the stable policy but it is fine since the module was added recently and it is less than likely that someone had already overwriten that method.

closes odoo/enterprise#4616

Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>