[FIX] account_reports: In Need of Action filter applied in wrong case
Steps to reproduce the bug:
- Create a 100€ invoice I for a customer C with a due date in the past
- Create a payment of 200€ for C
- Go to Followup reports

Bug:

The followup_status of C was 'no_action_needed' but the C was displayed with
the filter 'In Need of Action'.

Technically:

The function _compute_for_followup takes the overdue amount into account but the function
_get_partners_in_need_of_action doesn't.
The records to display with the filter 'In Need of Action' was only compted with
the function _get_partners_in_need_of_action. That's why there was a mismatch.

opw:1981905

closes odoo/enterprise#4617

Signed-off-by: Simon Goffin (sig) <sig@openerp.com>