[FIX] web_mobile: clear button on many2many_tags not working
Before this commit, the clear button was displayed when a many2many_tags
field was opened in the native app (Android/iOS).
But the behavior wasn't well defined for many2many fields as we can
have multiple values and a crash occured when you clicked on it.

We don't want to fix this by removing all values by clicking on
the clear button! We consider that it's safer to remove one by one
with the "close button" on each of them.
So we decided to hide this button for many2many fields and to keep
it for many2one fields.

Note that many2many_tags is oponed with the native many2oneDialog
because
this widget is implemented with a many2one.
In the future many2oneDialog should works with many2one and many2many.

opw-2008445

closes odoo/enterprise#4624

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>