[FIX] delivery_bpost: avoid wrong street number
In delivery bpost when the customer is in another country
than belgium, it doesn't exist a specific regex in order to
guess the street number. It will select the first decimal that could
be find. However it will fail with addresses that contains a district
number or something similar before it.

The steetNumber is not require and it's mainly used for us in order
to avoid to long streetName. On the label it will automaticaly be set
at the end of the streetName. So for each country we could define the
last word as the streetNumber in order to respect the initial given
format. However we only set the last word as a streetNumber if it
contains at least one decimal.