[FIX] account_reports: posted followup message
- In follow-up reports, open any customer follow-up
- Send a follow-up
- Edit the follow-up directly on the form

See the content of the already sent email changing accordingly to
modifications.

It happens because the posted message contains classes used for JS
interaction.

Removing the mentioned class when posting the message solves the
problem. Another solution would have been to change the JS selector.
It seems simpler and more robust to get rid of the classes.

opw-2024225

closes odoo/enterprise#4648

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>