[FIX] delivery_bpost: do not truncate addresses
Bpost only accept at the date of this commit:
- streetName with a maximum of 40 characters
- postalCode with a maximum of 8 characters
- locality with a maximum of 40 characters

In order to avoid an error, the system truncates the fields
if it contains more characters than allowed. However it will
create a label with a wrong or partial address that will fail
when the postman will try to deliver it and end in a lost or
a return. It's better to raise an error if the address is not
possible to communicate to bpost.