[FIX] account_reports: export xlsx file with custom date
Change the locale date format to replace the month number `%m` by the
abbr. of the month name `%b`. Export the partner ledger, traceback.

Some date appear formatted in the custom user's locale already, it is
not possible to parse those dates to a datetime object.

The fix writes the date as-is in the xlsx file.

opw-2026108

closes odoo/enterprise#4654

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>