[FIX] sale_subscription: missing email layout
1. Create a subscription templates with the Payment Mode to Invoice and
select an email template to be sent.
2. Select the template Invoice: Send by email.
3. Create a subscription, put it in progress and generate an invoice.

Before this commit, the email will be sent with all the information from
the template 'Send by email', but the final email lack the header, the
footer and all the relevant buttons for the customer to see their
invoice in the portal.

Now, the email will be sent out identical to the email sent with the
'Send & Print' button with the correct layout.

opw-2025099

closes odoo/enterprise#4661

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>