[FIX] l10n_mx_edi: remove cert->ok if no cert used
In 11.0, with d3c34de72 removing certificate used in an invoice was
prevented. But since the field l10n_mx_edi_cfdi_certificate_id is
computed non-stored we would have an error. With this change we only
prevent removing certificate if a certificate has been used.

In any case, the certificate can just be unlinked from current company
to prevent it being used.

opw-2029034
closes #4687

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>