[FIX] delivery_dhl: fixed the decimal precision value for dutiable good
when We send dutiable good to EU and USA. we got a error message from
DHL API that the value of the attribute [DeclaredValue] is wrong.
and DHL Express API requires a number with exact 2 digits after comma

So round the declared value to exactly 2 digits after comma.

opw-2024747 and 2030710