[FIX] web_enterprise: non visible content on iPad after navigation
On iPad when navigating from a view containing an onboarding banner the
content of the destination view doesn't appear.

This issue seems related to the scrolling ("overflow:auto") set on the
content area (".o_content") for medium and larger screens. For some
reason, Mobile Safari misses a repaint/redraw after the navigation has
occurred.

This commit fixes it by promoting the view controller container element
to a composite layer by applying a "translateZ()" transformation which
force a repaint.

The "-webkit" CSS prefix is voluntarily applied to narrow the target of
this rule to only Webkit/Blink browsers.

Note: Do not forward port this commit after 12.0 !

Reference: https://stackoverflow.com/a/21947628

opw-2006351

closes odoo/enterprise#4739

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>