[IMP] sale_ebay: new synchronisation based on orders
The eBay API intends some functions to be called often, some sparsely.
GetOrders is the recommended call to use for order (sales) management.
Contrast with GetItem, where the doc specifies: During the last 30 minutes
of an auction, the maximum frequency is every five minutes.
If the API is used incorrectly, its results are flaky and many errors
with Code: 10007 are returned by the server.

Note that this new synchronisation changes a few things:
 - Sale Orders are in bijection with eBay Orders, which may combine transactions
 - all orders are imported, including for products that have not been listed
 through Odoo. Such orders are not confirmed automatically.
 This is somewhat necessary since an order could contain some products listed
 via Odoo and some unlisted.
 - Taxes are taken directly from eBay and created on the fly if necessary
 - Upward synchronisation with eBay is done somewhat aynchronously using a queue
 (although the implementation is quite hackish).
 The idea is that since we can't synchronize based on the bijection between
 transactions and products since orders may combine multiple products, we need
 to have a system of the sort.

Another advantage of the new system is that if the synchronisation of some
orders failed for any reason, it can be retrieved again by a server action,
added as synchronize_orders_recovery.

Note that this synchronisation is not the default.
To use it, give an appropriate value to the ir.config_parameter 'ebay_last_sync'
(the date at which it should start basically).
Then remove the old eBay scheduled actions to replace them with the new one.

opw 1962525
opw 1924454