[FIX] web_studio: render report in studio as pdf
When some reports (for example, invoices) are open in studio to be
edited, they don't have the same look as when they are printed as pdf.
This issue arise because some reports display or hide parts depending on
if they are render in html or in pdf.

To avoid this discrepancy, the reports of type qweb-pdf will be rendered
as pdf in web studio.

opw-2032147

closes odoo/enterprise#4758

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>