[FIX] quality_control: access right issue
issue: When creating a SO with a user getting only the " sales User - Own document only"
right, he can't create a delivery order, the error shown is coming from Quality check.

Whether or not you are quality user, an xpath added fields regarding quality check and
so they were fetched, resulting in an access error. This commit disabled this xpath if
you're not quality user and now, no error appear (but obviously the user can only read
the picking, not process it).

opw-2006029

closes odoo/enterprise#4807

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>