[FIX] account_reports: Translation issue with overdue message
Steps to reproduce:

- Have follow-up level checked in General Settings->Accounting
- Have a customer with language different than English
- Have an invoice due for this customer for more than the follow-up level
- Go to follow-up report, text is in partner language
- Switch to manual, set date to tomorrow, click remind-me late

Bug:

The overdue message was in the lang of the company

PS:

In function _get_default_summary defined in model "account.followup.report",
the right translation was taken when a followup line was found by the function
get_followup_line (in module account_followup_report). In the other case,
the overdue message defined on the company of the current user was taken (in
module account_reports).

As the overdue message was not editable, it was not possible to customize it or
to translate it.

opw:1959216

closes odoo/enterprise#4762

Signed-off-by: Simon Goffin (sig) <sig@openerp.com>