[FIX] l10n_mx_edi performance in unistallation issue
When trying to uninstall the module l10n_mx_edi (which is installed
by default in all mexican databases) the uninstallation was taking
so long that it would never finish. This happened because of the
normal uninstallation procedure where the records are deleted one
by one (here, the problem is that the module creates more than 56000
records). As a result, the module could not be uninstalled and if
a customer would not like to use the SAT connection and signature
feature they would not be able to use the mexican localization at all.

Deleting the biggest batch of records (55000) with SQL amiliorates
a lot the performance.

opw 2025895 and opw 2032954

closes odoo/enterprise#4776

Signed-off-by: Richard Mathot (rim) <rim@openerp.com>