[FIX] account_reports: make select2 dynamic
Create tons of Analytic Accounts then go to Accounting > Reporting > US
GAAP > Profit and Loss. Filter the analytic by account, the rendering of
the options is very slow.

This change the behavior of the select2 widget by fixing a minimum input
length in the search bar to execute the selection. The length is based
on the records count so the more options are available the more precise
the filter will be.

See also 53606762

opw-2026863

closes odoo/enterprise#4771

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>