[FIX] web_studio: edit field in x2m view -> works
In 11.0 change 9c6cce81a4 used the correct x2m model if we are inside
an x2many view.

But the edition of selection field was added in saas-11.2 (with 8190e17)
and the change has not been made for them, so if the model of the x2m is
different that the parent view model, it would possibly not work or work
wrongly.

Without fix, added test failed first assertion (of 3) with:

{expected: "product", result: "coucou"}

opw-2034278
closes #4793

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>