[FIX] delivery_easypost: HS code for Easypost.
Because of odoo/2d738f8563212fbdf10924468b1fcd865524e6d7,
the hs_code field was moved to a separate module.
This module is temporary since the field has been moved back in
delivery in master.
It is in auto_install, but it is not a dependency of delivery_easypost.
However, we need the hs_code in easypost since this is a required field
by some delivery providers.
Therefore we get it only if the module is installed,
and otherwise we send dummy data (which should be enough if not required).

opw 2031249

closes odoo/enterprise#4799

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>