[FIX] web_studio: split report document to allow translations
Before this commit, when creating a new report in web studio, the user
can't translate the report.

This commit changes the structure of new reports to allow the
possibility to translate the report. The new structure is based on :
https://github.com/odoo/odoo/blob/12.0/doc/reference/reports.rst#translatable-templates

Indeed, to create a translatable report, two templates must be defined.
The main report template must call the translatable document template
with the attribute t-lang set to a language code, thanks to this
attribute the document template will be translated.

opw-2031837

closes odoo/enterprise#4827

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>