[FIX] account_sepa: missing creditor's address in sepa credit transfer file
Before this commit, the generated sepa credit transfer file is missing
the address information of the creditor. Even if those informations are
not required, they are recommended as it can lead to a rejection of the
payment.

Now, the creditor sepa transfer file add the address information of the
creditor.

opw-2031912

closes odoo/enterprise#4871

Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>