[FIX] ocn_client: avoid to notify deleted account
Before this commit there was an issue when you had two or
more accounts in the mobile app.
After removing an account, the app tried to synchronise accounts
on OCN server but the app didn't store OCN_TOKEN for users.
Therefore OCN server didn't do anything during the sync as we
can't compare tokens present in the app and the ones on OCN server
to know wich one to delete.

After this commit ocn_client will return OCNToken for each users.
So, the app will store it and will be able to use it during
the synchronisation.

This feature was already implemented in saas-12.4 See:
odoo/enterprise@9ebed6d9318b41dc53c3b043f22335a40e6bb7db
We need to backport this as a fix.

Steps to reproduce:
- install Odoo App 2.2.1+
- login with User 1
- login with User 2
- now delete User 2
- close Odoo App
- now send notification to User 2 from User 1
- you get notification for User 2, but User 2 is deleted. (BUG)

opw-2038331

closes odoo/enterprise#4876

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>