[REF] account_reports: remove dead code
Since commit 833991e489324d69a262168fd53522ff26353436, the variables
`where_clause` and `where_params` are not used any more in the cash flow
statement query.
`tables` was still used; however, as `sub_where_clause` was used with
it, it makes more sense to use `sub_tables`.

Thanks to Laurent Smet (las) for spotting this with me!

closes odoo/enterprise#4887

Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>