[FIX] account_reports: fix journal filter cash flow statement
When filtering by a journal on the cash flow statement, all the move
lines were filtered on this journal.
However, for all the report lines other than CASHSTART and CASHEND, we
need to include the reconciled entries too, which may be on a different
journal.

This fix removes the journals filter for most outer queries of the cash
flow statement, and leaves it on the innermost WITH statements.

To reproduce the problem on the runbot:

1.  register the payment of an invoice with the default data
2.  go to the cash flow statement
3.  filter on the "Bank" journal
4.  see an unexplained difference

With this PR, there is no unexplained difference any more, because all
the account move lines reconciled with an account move line from the
selected journal are included in the report.

closes odoo/enterprise#4891

Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>