[FIX] account_invoice_extract: pre-update OCR status
Clicking on the 'Update status' button of an invoice processed by the
OCR will now update the status of the current pending OCR request, but
also the status for a few other pending OCR requests. The goal is to
improve the UX by reducing the number of times we have to manually
click on that button.

A cron is also added to update those statuses periodically.

closes odoo/enterprise#4796

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>