[FIX] stock_account_enterprise: incorrect incoming/outgoing inventory in warehouse dashboard
Before this commit, when a product had its category setted to automated
and FIFO, the warehouse dashboard in the inventory gave incorrect values
for the incoming/outgoing inventory.

This arises because :
- The product's categories with more than one property, generate
multiple lines in the SQL request therefore it will add multiple times
the same move. To create the report only the 'property_cost_method'
property is needed;
- For the valuation of the outgoing inventory, the products with cost
methods 'fifo' and 'average' gives a negative valuation; in contrary,
the products with 'standard' as cost method gives always a positive
valuation. The valuation needs to be always positive, to avoid problems
when mixing products with different cost methods.

Now, the values in the dashboard are correct, independently of the
category of the product.

opw-2029912

closes odoo/enterprise#4918

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>