[FIX] web_gantt: task name column resize
Before this commit, when the task name was too big, the text was
truncated without ellipses, and there were not possibility to resize the
column to see the complete text.

Now, the first column can be resized, and the text is truncated with
ellipses to show the user that the text is bigger.

opw-2027249

closes odoo/enterprise#4925

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>