[FIX] stock_barcode: package in inventory adjustment
In the inventory adjustment, if we scan a package the information is
not transferred in many instances and so we get js traceback or
information loss.

For example: scan a package, go to edit => js traceback

With this changeset, we:

- keep information of package through steps
- do not set default product quantity to the one of last scanned package
- do not display button "Put In Pack" in inventory adjustment
- display package on inventory adjustment lines list
- display an error when editing a line from wrong data

Without the change, the added test fails when editing the line because
things do not work after traceback "Cannot read property 'top' of
undefined" inside method "_highlightLine" of LinesWidget.

opw-2041525
opw-2046224
closes #4995
closes #4931

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>