[FIX] website_enterprise: restore design of go-to-backend button
The design of the button (when the user clicks on it) was broken by
https://github.com/odoo/odoo/commit/332d35b56e67119ce55f3260501227ad9b84ba14

Indeed the CSS class should never have been removed. This commit solves
the issue by restoring the class and solves the original goal of the
mentioned commit by redirecting to backend only if the button is not
already in its loading state.

task-2041973

closes odoo/enterprise#4962

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>


Co-authored-by: Jérémy Kersten (jke) <jke@openerp.com>
Co-authored-by: qsm-odoo <qsm@odoo.com>