[FIX] l10n_nl_reports: XAF report with long login
Change your login so it has strictly more than 35 characters, do a
reconciliation and export the general ledger. Invalid User Error because
the `/company/customersSuppliers/changeInfo/userID` must be at most 35
characters long.

The related field contains "the identifier of the user who did the
latest modifications". That field can be any kind of identifier. The fix
introduces a logic to select the best available identifier. First the
login, then the display name and finally (if none of the two first are
less than 35 characters) the truncated display name followed by the
internal id.

opw-2038376

closes odoo/enterprise#4946

Signed-off-by: Julien Castiaux <Julien00859@users.noreply.github.com>